+++ packages/pypi/anywhere_agents/__init__.py
@@ -2,4 +2,4 @@
 
 See https://github.com/yzhao062/anywhere-agents for details.
 """
-__version__ = "0.2.0"
+__version__ = "0.3.0"